requirements.txt
aiofiles==24.1.0
cffi==1.17.0
cryptography==43.0.0
greenlet==3.0.3
html5tagger==1.3.0
httptools==0.6.1
jwt==1.3.1
multidict==6.0.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
sanic==24.6.0
sanic-ext==23.12.0
sanic-jwt==1.8.0
sanic-openapi==21.12.0
sanic-routing==23.12.0
SQLAlchemy==2.0.32
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
uvloop==0.19.0
websockets==12.0
